+++ python/rigol_fastrec/__init__.py
@@ -28,7 +28,7 @@ from .scpi import ScpiControl
 # logging config.
 logging.getLogger("rigol_fastrec").addHandler(logging.NullHandler())
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = [
     "WaveRecorder",